Get test db functioning
Uses fixtures to create the test db
