+++ config.py
@@ -1,6 +1,5 @@
 from os import environ
 
-
 class Config:
     # TESTING = environ.get('TESTING')
     # FLASK_DEBUG = environ.get('FLASK_DEBUG')
