+++ conftest.py
@@ -1,0 +1,1 @@
+from sabelotodo.tests.test_fixtures import database, app, _db
