+++ sabelotodo/models.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from . import db
+from sabelotodo import db
 
 
 @dataclass
